% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw USA Map and add information on accidents by depicting their location on the map
its longitude and lattitude coordonates}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer: represents the code number of a state in USA and takes it value between 1 to 56.}

\item{year}{integer: defines the indicated year whose information on accidents is needed.}
}
\value{
Map or graphic object.
}
\description{
Draw USA Map and add information on accidents by depicting their location on the map
its longitude and lattitude coordonates
}
\examples{
\dontrun{
fars_map_state(state.num = 2, year = 2013)
}
}
\seealso{
\itemize{
 \item \code{\link{make_filename}} to create data file name in function of year
 \item \code{\link{fars_read}} to read csv file in order to produce a rectangular table
}
}
