% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file if it exist in table format and creates a data frame from it}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string and defines the name of the file where the data is going to be read from.
filename represents also the path to get file.}
}
\value{
A data frame: a type of data structure used by R to represents rectangular data for statistical analysis.
However if the filename doesn't exist the \code{fars_read} is stopped with error message.
}
\description{
Read csv file if it exist in table format and creates a data frame from it
}
\examples{
\dontrun{
mydataInfolderOther <- fars_read(filename = "folder_data/my_data.csv")
}
\dontrun{
mydataInfolderCurrent <- fars_read(filename = "my_data.csv")
}
}
