% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create to a list of years and months for each year in years
Create filename for a given list vector of year and for each year the function
return a data frame with month in which at least an accident is realized for each year in years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{list or vector of integer
which contains the years for which the accidents data are needed}
}
\value{
A list or data.frame of the same length as \code{years} which contains
 for each year from \code{years} the list of Months of each accident.
 if in the list of \code{years} a year is not numeric [fars_read_years]
 returns NULL with warning message.
}
\description{
Create to a list of years and months for each year in years
Create filename for a given list vector of year and for each year the function
return a data frame with month in which at least an accident is realized for each year in years
}
\examples{
\dontrun{
list_filename <- fars_read_years(years = "2013")
}
\dontrun{
list_filename <- fars_read_years(years = c("2013", "2014", "2015"))
}
}
\seealso{
\itemize{
* \code{\link{make_filename}} to create filename
* \code{\link{fars_read}} to read accidents data from csv format
}
}
