% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create Cross table to Compare the couple of year and month according to the number of accidents
This function propose a contingence table whose row is year and column represents the months in which
the number of accidents is derived at each interception.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{list or vector of concerned years}
}
\value{
tibble or dataframe of Cross table between \code{years} in row and months in column
which contains the number of accidents.
}
\description{
Create Cross table to Compare the couple of year and month according to the number of accidents
This function propose a contingence table whose row is year and column represents the months in which
the number of accidents is derived at each interception.
}
\seealso{
\code{\link{fars_read_years}} to create an accident dataframe according the months and year accidents
}
