% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create character string corresponding to filename or data base name for given input year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer: represents the year for which the data on accidents are needed}
}
\value{
Character string formatted such as accident_year.csv.bz2.
}
\description{
Create character string corresponding to filename or data base name for given input year
}
\examples{
\dontrun{
filename <- make_filename(year = 2013)
}
\dontrun{
filename <- make_filename(year = 2015)
}
}
